requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
graphql-core==3.2.7
idna==3.11
lia-web==0.2.3
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
responses==0.25.8
schema==0.7.8
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.285.0
strawberry-graphql-django==0.67.0
typing_extensions==4.15.0
urllib3==2.5.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Testing dependencies
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-bdd==7.0.1
factory-boy==3.3.0
faker==22.6.0
freezegun==1.4.0
